Migrate C runtime to 64-bit
- runtime.c: remove -m32 from JIT_COMPILE_COMMAND, change
  function pointer typedef to long (*)(long*, long), all
  stack/bytecode parameters from int to long, add guard_fail_arr
  for per-PC guard failure tracking
- runtime.h: update function signatures to use long
- runtime_camlwrap.c/h: int -> long, Val_int -> Val_long,
  cast function pointers for caml_alloc_array
- interop_caml.c: int -> long, Val_int -> Val_long,
  Int_val -> Long_val
- main.c: divide/modulo functions use long
- test/tlbc_example/stub.c: same int -> long changes
- Remove caml/compatibility.h includes (removed in OCaml 5.x)

+++ runtime/main.c
@@ -21,9 +21,9 @@ unsigned long get_current_micros() {
   return current_time.tv_sec * (int)1e6 + current_time.tv_usec;
 }
 
-int divide(int rhs, int lhs) { return rhs / lhs; }
+long divide(long rhs, long lhs) { return rhs / lhs; }
 
-int modulo(int rhs, int lhs) { return rhs % lhs; }
+long modulo(long rhs, long lhs) { return rhs % lhs; }
 
 /* "stderr" is a macro and cannot be referred to in libmincaml.S, so */
 /*    this "min_caml_stderr" is used (in place of "__iob+32") for better */
