+++ runtime/runtime.h
@@ -2,6 +2,7 @@
 #define RUNTIME_H_
 
 #include <stdbool.h>
+#include <stdint.h>
 
 enum jit_mode { NORMAL, HYBRID_TJ, HYBRID_MJ, HYBRID_ALL };
 
@@ -9,10 +10,10 @@ extern bool no_jit;
 
 void set_jit_mode(enum jit_mode);
 
-void c_can_enter_jit(int *, int, int *, int);
+void c_can_enter_jit(long *, long, long *, long);
 
-void c_jit_setup(int *, int, int *, int);
+void c_jit_setup(long *, long, long *, long);
 
-void c_jit_merge_point(int*, int, int*, int);
+void c_jit_merge_point(long*, long, long*, long);
 
 #endif
